Moved the differentiating functions to their own submodule
